// src/app/core/models/ms-accounting/mand-expensecont.ts
export interface IMandExpenseCont {
  spentId: string;
  mandxexpensecontId: string;
  cvman: string;
  amount: string;
  cabms: string;
  vat: string;
  departure: string;
  cooperation: string;
  descabms: string;
  retentionisr: string;
  retentionvat: string;
  total: string;
  categorycabms: string;
  appliesto?: any;
  departurestop: string;
}